+++ src/elements/Order.php
@@ -15,6 +15,7 @@ use craft\commerce\base\GatewayInterface;
 use craft\commerce\base\OrderDeprecatedTrait;
 use craft\commerce\base\OrderValidatorsTrait;
 use craft\commerce\base\ShippingMethodInterface;
+use craft\commerce\db\Table;
 use craft\commerce\elements\actions\UpdateOrderStatus;
 use craft\commerce\elements\db\OrderQuery;
 use craft\commerce\errors\OrderStatusException;
@@ -671,7 +672,7 @@ class Order extends Element
         // Try to catch where the order could be marked as completed twice at the same time, and thus cause a race condition.
         $completedInDb = (new Query())
             ->select('id')
-            ->from(['{{%commerce_orders}}'])
+            ->from([Table::ORDERS])
             ->where(['isCompleted' => true])
             ->andWhere(['id' => $this->id])
             ->exists();
@@ -680,9 +681,6 @@ class Order extends Element
             $mutex->release($lockName);
             return true;
         }
-        // Release after we have confirmed this order is not already complete
-
-        $mutex->release($lockName);
 
         $this->isCompleted = true;
         $this->dateOrdered = Db::prepareDateForDb(new DateTime());
@@ -709,17 +707,20 @@ class Order extends Element
             $this->trigger(self::EVENT_BEFORE_COMPLETE_ORDER);
         }
 
-        if (Craft::$app->getElements()->saveElement($this, false)) {
+        $success = Craft::$app->getElements()->saveElement($this, false);
 
-            $this->afterOrderComplete();
+        $mutex->release($lockName);
 
-            return true;
+        if (!$success) {
+            Craft::error(Craft::t('commerce', 'Could not mark order {number} as complete. Order save failed during order completion with errors: {order}',
+                ['number' => $this->number, 'order' => json_encode($this->errors)]), __METHOD__);
+
+            return false;
         }
 
-        Craft::error(Craft::t('commerce', 'Could not mark order {number} as complete. Order save failed during order completion with errors: {order}',
-            ['number' => $this->number, 'order' => json_encode($this->errors)]), __METHOD__);
+        $this->afterOrderComplete();
 
-        return false;
+        return true;
     }
 
     /**
@@ -1923,7 +1924,7 @@ class Order extends Element
 
             $count = (new Query())
                 ->where(['o.orderStatusId' => $orderStatus->id, 'e.dateDeleted' => null])
-                ->from(['{{%commerce_orders}} o'])
+                ->from([Table::ORDERS . ' o'])
                 ->leftJoin(['{{%elements}} e'], '[[o.id]] = [[e.id]]')
                 ->count();
 
@@ -1933,17 +1934,16 @@ class Order extends Element
                 'label' => $orderStatus->name,
                 'criteria' => $criteriaStatus,
                 'defaultSort' => ['dateOrdered', 'desc'],
-                'badgeCount' => $count
+                'badgeCount' => $count,
+                'data' => [
+                    'handle' => $orderStatus->handle
+                ]
             ];
         }
 
         $sources[] = ['heading' => Craft::t('commerce', 'Carts')];
 
-        $edge = new DateTime();
-        $interval = new DateInterval('PT1H');
-        $interval->invert = 1;
-        $edge->add($interval);
-        $edge = $edge->format(DateTime::ATOM);
+        $edge = Plugin::getInstance()->getCarts()->getActiveCartEdgeDuration();
 
         $updatedAfter = [];
         $updatedAfter[] = '>= ' . $edge;
@@ -1954,6 +1954,9 @@ class Order extends Element
             'label' => Craft::t('commerce', 'Active Carts'),
             'criteria' => $criteriaActive,
             'defaultSort' => ['commerce_orders.dateUpdated', 'asc'],
+            'data' => [
+                'handle' => 'cartsActive'
+            ]
         ];
         $updatedBefore = [];
         $updatedBefore[] = '< ' . $edge;
@@ -1963,7 +1966,10 @@ class Order extends Element
             'key' => 'carts:inactive',
             'label' => Craft::t('commerce', 'Inactive Carts'),
             'criteria' => $criteriaInactive,
-            'defaultSort' => ['commerce_orders.dateUpdated', 'desc']
+            'defaultSort' => ['commerce_orders.dateUpdated', 'desc'],
+            'data' => [
+                'handle' => 'cartsInactive'
+            ]
         ];
 
         $criteriaAttemptedPayment = ['hasTransactions' => true, 'isCompleted' => 'not 1'];
@@ -1972,6 +1978,9 @@ class Order extends Element
             'label' => Craft::t('commerce', 'Attempted Payments'),
             'criteria' => $criteriaAttemptedPayment,
             'defaultSort' => ['commerce_orders.dateUpdated', 'desc'],
+            'data' => [
+                'handle' => 'cartsAttemptedPayment'
+            ]
         ];
 
         return $sources;