+++ src/records/OrderHistory.php
@@ -7,6 +7,7 @@
 
 namespace craft\commerce\records;
 
+use craft\commerce\db\Table;
 use craft\db\ActiveRecord;
 use DateTime;
 use yii\db\ActiveQueryInterface;
@@ -38,7 +39,7 @@ class OrderHistory extends ActiveRecord
      */
     public static function tableName(): string
     {
-        return '{{%commerce_orderhistories}}';
+        return Table::ORDERHISTORIES;
     }
 
     /**