+++ src/records/PaymentCurrency.php
@@ -7,6 +7,7 @@
 
 namespace craft\commerce\records;
 
+use craft\commerce\db\Table;
 use craft\db\ActiveRecord;
 
 /**
@@ -29,6 +30,6 @@ class PaymentCurrency extends ActiveRecord
      */
     public static function tableName(): string
     {
-        return '{{%commerce_paymentcurrencies}}';
+        return Table::PAYMENTCURRENCIES;
     }
 }