+++ src/records/Product.php
@@ -7,6 +7,7 @@
 
 namespace craft\commerce\records;
 
+use craft\commerce\db\Table;
 use craft\db\ActiveRecord;
 use craft\records\Element;
 use DateTime;
@@ -49,7 +50,7 @@ class Product extends ActiveRecord
      */
     public static function tableName(): string
     {
-        return '{{%commerce_products}}';
+        return Table::PRODUCTS;
     }
 
     /**