+++ src/records/SaleCategory.php
@@ -7,6 +7,7 @@
 
 namespace craft\commerce\records;
 
+use craft\commerce\db\Table;
 use craft\db\ActiveRecord;
 use craft\records\Category;
 use yii\db\ActiveQueryInterface;
@@ -32,7 +33,7 @@ class SaleCategory extends ActiveRecord
      */
     public static function tableName(): string
     {
-        return '{{%commerce_sale_categories}}';
+        return Table::SALE_CATEGORIES;
     }
 
     /**