+++ src/records/SalePurchasable.php
@@ -7,6 +7,7 @@
 
 namespace craft\commerce\records;
 
+use craft\commerce\db\Table;
 use craft\db\ActiveRecord;
 use yii\db\ActiveQueryInterface;
 
@@ -32,7 +33,7 @@ class SalePurchasable extends ActiveRecord
      */
     public static function tableName(): string
     {
-        return '{{%commerce_sale_purchasables}}';
+        return Table::SALE_PURCHASABLES;
     }
 
     /**