+++ src/records/SaleUserGroup.php
@@ -7,6 +7,7 @@
 
 namespace craft\commerce\records;
 
+use craft\commerce\db\Table;
 use craft\db\ActiveRecord;
 use craft\records\UserGroup;
 use yii\db\ActiveQueryInterface;
@@ -32,7 +33,7 @@ class SaleUserGroup extends ActiveRecord
      */
     public static function tableName(): string
     {
-        return '{{%commerce_sale_usergroups}}';
+        return Table::SALE_USERGROUPS;
     }
 
     /**