+++ src/records/Subscription.php
@@ -7,6 +7,7 @@
 
 namespace craft\commerce\records;
 
+use craft\commerce\db\Table;
 use craft\db\ActiveRecord;
 use craft\records\User;
 use DateTime;
@@ -44,7 +45,7 @@ class Subscription extends ActiveRecord
      */
     public static function tableName(): string
     {
-        return '{{%commerce_subscriptions}}';
+        return Table::SUBSCRIPTIONS;
     }
 
     /**