+++ src/services/Gateways.php
@@ -11,9 +11,11 @@ use Craft;
 use craft\commerce\base\Gateway;
 use craft\commerce\base\GatewayInterface;
 use craft\commerce\base\SubscriptionGateway;
+use craft\commerce\db\Table;
 use craft\commerce\gateways\Dummy;
 use craft\commerce\gateways\Manual;
 use craft\commerce\gateways\MissingGateway;
+use craft\commerce\Plugin;
 use craft\commerce\records\Gateway as GatewayRecord;
 use craft\db\Query;
 use craft\errors\MissingComponentException;
@@ -237,7 +239,7 @@ class Gateways extends Component
         $existingGateway = $this->getGatewayByHandle($gateway->handle);
 
         if ($existingGateway && (!$gateway->id || $gateway->id != $existingGateway->id)) {
-            $gateway->addError('handle', Craft::t('commerce', 'That handle is already in use.'));
+            $gateway->addError('handle', Plugin::t( 'That handle is already in use.'));
             return false;
         }
 
@@ -261,7 +263,7 @@ class Gateways extends Component
         $projectConfig->set($configPath, $configData);
 
         if ($isNewGateway) {
-            $gateway->id = Db::idByUid('{{%commerce_gateways}}', $gatewayUid);
+            $gateway->id = Db::idByUid(Table::GATEWAYS, $gatewayUid);
         }
 
         return true;
@@ -342,7 +344,7 @@ class Gateways extends Component
     {
         $projectConfig = Craft::$app->getProjectConfig();
 
-        $uidsByIds = Db::uidsByIds('{{%commerce_gateways}}', $ids);
+        $uidsByIds = Db::uidsByIds(Table::GATEWAYS, $ids);
 
         foreach ($ids as $gatewayOrder => $gatewayId) {
             if (!empty($uidsByIds[$gatewayId])) {
@@ -436,7 +438,7 @@ class Gateways extends Component
                 'uid',
                 'sortOrder'
             ])
-            ->from(['{{%commerce_gateways}}']);
+            ->from([Table::GATEWAYS]);
     }
 
     /**